Testing work on the FOV system
I started working through a tutorial to set up field of view for the players. I got the FOV mesh to draw properly but it's not actually stopping when it hits a sprite. I've put a comment on the video I was watching in hopes someone gets back to me.