Claim Menu - Drag and Drop
Started working on the functionality to drag and drop the claim coins onto the player icon in the Claim Menu.

I set it up so the ME coin can be drug and attach to the purple player placeholder text.

Current problem: The ME coin snaps off to the side for some reason rather than to the middle of the placeholder object.